refactor: nmeaparse; externalize util methods
